Enable do-notation when writing programs in the Program data type.
I've opted for a minimally invasive way of implementing this. The Program
data type is completely untouched and so there is very little that needs to
change in other modules.

There is a new P monad which enables writing things in do-notation. When
running P using the runP function a Program tree will be produced. There are
functions in the P monad which correspond to the constructors in the Program
data type, such as 'forAll', 'allocate', 'assign' and so forth.

The implementation of the P monad is rather subtle because it needs to
generate fresh names on the go. Normally, when providing a monadic interface
one can leave that to the code generator by including a bind constructor,
corresponding to the (>>=) operator in the monad class. Since this constructor
is higher order it doesn't have to worry about names. However, since I opted
to go for a minimally invasive change I didn't introduce a bind constructor.
Names are only needed in one place, and that is for the atomic operations. I
figured that it would be too big of a change to rewrite big parts of Obsidian
just to allow for atomic operations. Hence I decided to do the name generation
in the P monad. It uses a NameSupply monad which provides fresh names. That
monad in turn uses value supplies. At first I used a state monad to pass
around a state but that turned out not to work for functions like 'forAll'
where the monad needs to be run inside of a function. Using the trick with
passing a splittable name supply solved the problem.

Examples of using the do-notation:
runP (forAll 32 (\i -> do {p <- atomicOp "a" i AtomicInc; assign "b" 0 p}))

runP (do { assign "a" 32 (100 :: Data Word32); assign "b" 10 (40 :: Data Word32); a <- atomicOp "c" 0 AtomicInc; assign "d" 0 a})